+++ src/outputUtils.ts
@@ -4,6 +4,7 @@ import * as vscode from 'vscode';
 import { getConfig } from './dynamicConfig';
 import { IsWindows } from './checkTool';
 import { replaceTextByRegex } from './utils';
+import { Z_MEM_ERROR } from 'zlib';
 
 export const RunCmdTerminalName = 'MSR-RUN-CMD';
 const OutputChannelName = 'MSR-Def-Ref';
@@ -15,6 +16,33 @@ const ShellPath = IsWindows ? WindowsShell : 'bash';
 const ClearCmd = IsWindows && !UsePowershell ? 'cls' : "clear";
 
 const ShowColorHideCmdRegex = /\s+-[Cc](\s+|$)/g;
+
+export enum MessageLevel {
+	DEBUG = 0,
+	INFO = 1,
+	WARN = 2,
+	ERROR = 3,
+	FATAL = 4
+}
+
+export function outputMessage(level: MessageLevel, message: string, showWindow: boolean = true) {
+	switch (level) {
+		case MessageLevel.DEBUG:
+			outputDebug(message, showWindow);
+			break;
+		case MessageLevel.INFO:
+			outputInfo(message, showWindow);
+			break;
+		case MessageLevel.WARN:
+			outputWarn(message, showWindow);
+			break;
+		case MessageLevel.ERROR:
+		case MessageLevel.FATAL:
+		default:
+			outputError(message, showWindow);
+			break;
+	}
+}
 
 let _channel: vscode.OutputChannel;
 let _terminal: vscode.Terminal | undefined;
@@ -31,43 +59,53 @@ export function disposeTerminal() {
 	_terminal = undefined;
 }
 
-export function runCommandInTerminal(cmd: string, mustShowTerminal: boolean = false) {
+export function runCommandInTerminal(cmd: string, mustShowTerminal: boolean = false, clearTerminalAtFirst = true) {
 	cmd = enableColorAndHideCommandline(cmd);
 	// cmd += ' -M '; // to hide summary.
 	showTerminal(mustShowTerminal);
-	//vscode.commands.executeCommand('workbench.action.terminal.clear');
-	getTerminal().sendText(ClearCmd);
+	if (clearTerminalAtFirst) {
+		// vscode.commands.executeCommand('workbench.action.terminal.clear');
+		getTerminal().sendText(ClearCmd);
+	} else {
+		getTerminal().sendText('\n');
+	}
+
 	getTerminal().sendText(cmd);
 	showTerminal(mustShowTerminal);
 }
 
-export function outputWarn(message: string) {
+export function outputWarn(message: string, showWindow: boolean = true) {
 	getOutputChannel().appendLine(message);
-	showOutputChannel();
+	showOutputChannel(showWindow);
 }
 
-export function outputError(message: string) {
+export function outputError(message: string, showWindow: boolean = true) {
 	getOutputChannel().appendLine(message);
-	showOutputChannel();
+	showOutputChannel(showWindow);
 }
 
-export function outputResult(text: string) {
+export function outputResult(text: string, showWindow: boolean = true) {
 	getOutputChannel().appendLine(text);
-	showOutputChannel();
+	showOutputChannel(showWindow);
 }
 
-export function outputInfo(message: string) {
+export function outputKeyInfo(text: string) {
+	getOutputChannel().appendLine(text);
+	showOutputChannel(true, true);
+}
+
+export function outputInfo(message: string, showWindow: boolean = true) {
 	if (getConfig().ShowInfo && !getConfig().IsQuiet) {
 		getOutputChannel().appendLine(message);
-		showOutputChannel();
+		showOutputChannel(showWindow);
 	}
 }
 
-export function outputDebugOrInfo(isDebug: boolean, message: string) {
+export function outputDebugOrInfo(isDebug: boolean, message: string, showWindow: boolean = true) {
 	if (isDebug) {
-		outputDebug(message);
+		outputDebug(message, showWindow);
 	} else {
-		outputInfo(message);
+		outputInfo(message, showWindow);
 	}
 }
 
@@ -82,8 +120,13 @@ export function clearOutputChannel() {
 	getOutputChannel().clear();
 }
 
-export function enableColorAndHideCommandline(cmd: string): string {
-	return replaceTextByRegex(cmd, ShowColorHideCmdRegex, '$1');
+export function enableColorAndHideCommandline(cmd: string, removeSearchWordHint: boolean = true): string {
+	let text = replaceTextByRegex(cmd, ShowColorHideCmdRegex, '$1');
+	if (removeSearchWordHint) {
+		text = replaceTextByRegex(text, /\s+Search\s+%~?1[\s\w]*/, ' ');
+	}
+
+	return text.replace(/\s+Search\s*$/, '');
 }
 
 function showTerminal(mustShowTerminal: boolean = false) {
@@ -92,8 +135,8 @@ function showTerminal(mustShowTerminal: boolean = false) {
 	}
 }
 
-function showOutputChannel(showWindow: boolean = true) {
-	if (showWindow && !getConfig().IsQuiet) {
+function showOutputChannel(showWindow: boolean = true, ignoreQuiet: boolean = false) {
+	if (showWindow && (ignoreQuiet || !getConfig().IsQuiet)) {
 		getOutputChannel().show(true);
 	}
 }